Pass comparison callback to sort radix-sort.test.js
**unit tests bug**
unit tests rely on `Array.sort` . However, there is no comparison function which will sort the array as they were strings. 

Not sure why tests passed in the case of the solution file. Sandbox is not very good for debugging but test should also fail in there bc of the bug in the test not the algo